refactor: split oversized files into focused modules across the app
Extract large components, hooks, and Rust modules into submodules to keep
files under the 600-line limit (recorder timeline, side panels, renderer,
TTS/ASR runtimes, overlays, third-party ASR crates). Also finalize the
zoom-block timeline UI: badge centered on the solid hold range, a
portal-rendered float badge above narrow blocks, and simplified
transition/edge handles.

+++ src/overlay/text_input/passive.rs
@@ -259,7 +259,7 @@ fn translate_vk(vk_code: u32) -> Option<String> {
     }
 
     let layout = unsafe { GetKeyboardLayout(0) };
-    let scan_code = unsafe { MapVirtualKeyW(vk_code, MAPVK_VK_TO_VSC) as u32 };
+    let scan_code = unsafe { MapVirtualKeyW(vk_code, MAPVK_VK_TO_VSC) };
     let mut utf16 = [0u16; 8];
     let translated = unsafe {
         ToUnicodeEx(
